+++ SnakeG/Model/Auth.swift
@@ -1,8 +1,8 @@
 //
-//  GameView.swift
+//  Auth.swift
 //  SnakeG
 //
-//  Created by Noujan Fakhri on 10/7/21.
+//  Created by Noujan Fakhri on 11/28/21.
 //
 
 import Foundation
@@ -55,34 +55,4 @@ class AppViewModel: ObservableObject {
         self.signedIn = false
     }
     
-}
-
-struct GameView : View {
-    
-    @StateObject var viewModel = AppViewModel()
-    
-    var body: some View {
-        ZStack{
-            if viewModel.signedIn {
-                SnakeGameView()
-            } else {
-                SignUpView()
-            }
-        }
-        .environmentObject(viewModel)
-        .onAppear() {
-            viewModel.signedIn = viewModel.isSignedIn
-        }
-    }
-}
-
-
-struct GameView_Previews: PreviewProvider {
-    static var previews: some View {
-        Group {
-            GameView()
-                .previewDevice(PreviewDevice(rawValue: "iPhone 12 Pro"))
-        }
-        
-    }
 }
