--- SnakeG/session.swift
@@ -1,11 +1,0 @@
-//
-//  session.swift
-//  SnakeG
-//
-//  Created by Noujan Fakhri on 4/25/21.
-//
-
-import Foundation
-import Firebase
-
-
